Moved to pipeline utility functions to compute distances and radius of gyration.
These functiona are used in different modules of the codebase.
