+++ Yank/tests/test_pipeline.py
@@ -18,12 +18,30 @@ import os
 from simtk import openmm
 
 from yank import utils
-from yank.pipeline import find_components
+from yank.pipeline import *
 
 
 # =============================================================================
 # TESTS
 # =============================================================================
+
+def test_compute_min_dist():
+    """Test computation of minimum distance between two molecules"""
+    mol1_pos = np.array([[-1, -1, -1], [1, 1, 1]], np.float)
+    mol2_pos = np.array([[3, 3, 3], [3, 4, 5]], np.float)
+    mol3_pos = np.array([[2, 2, 2], [2, 4, 5]], np.float)
+    assert compute_min_dist(mol1_pos, mol2_pos, mol3_pos) == np.sqrt(3)
+
+
+def test_compute_min_max_dist():
+    """Test compute_min_max_dist() function."""
+    mol1_pos = np.array([[-1, -1, -1], [1, 1, 1]])
+    mol2_pos = np.array([[2, 2, 2], [2, 4, 5]])  # determine min dist
+    mol3_pos = np.array([[3, 3, 3], [3, 4, 5]])  # determine max dist
+    min_dist, max_dist = compute_min_max_dist(mol1_pos, mol2_pos, mol3_pos)
+    assert min_dist == np.linalg.norm(mol1_pos[1] - mol2_pos[0])
+    assert max_dist == np.linalg.norm(mol1_pos[1] - mol3_pos[1])
+
 
 def test_method_find_components():
     """Test find_components() function."""
