+++ Yank/tests/test_yaml.py
@@ -177,37 +177,19 @@ def get_template_script(output_dir='.'):
 
     return yank_load(template_script)
 
+
 # ==============================================================================
 # YamlBuild utility functions
 # ==============================================================================
-
-
-def test_compute_min_dist():
-    """Test computation of minimum distance between two molecules"""
-    mol1_pos = np.array([[-1, -1, -1], [1, 1, 1]], np.float)
-    mol2_pos = np.array([[3, 3, 3], [3, 4, 5]], np.float)
-    mol3_pos = np.array([[2, 2, 2], [2, 4, 5]], np.float)
-    assert compute_min_dist(mol1_pos, mol2_pos, mol3_pos) == np.sqrt(3)
-
-
-def test_compute_min_max_dist():
-    """Test compute_min_max_dist() function."""
-    mol1_pos = np.array([[-1, -1, -1], [1, 1, 1]])
-    mol2_pos = np.array([[2, 2, 2], [2, 4, 5]])  # determine min dist
-    mol3_pos = np.array([[3, 3, 3], [3, 4, 5]])  # determine max dist
-    min_dist, max_dist = compute_min_max_dist(mol1_pos, mol2_pos, mol3_pos)
-    assert min_dist == np.linalg.norm(mol1_pos[1] - mol2_pos[0])
-    assert max_dist == np.linalg.norm(mol1_pos[1] - mol3_pos[1])
-
 
 def test_remove_overlap():
     """Test function remove_overlap()."""
     mol1_pos = np.array([[-1, -1, -1], [1, 1, 1]], np.float)
     mol2_pos = np.array([[1, 1, 1], [3, 4, 5]], np.float)
     mol3_pos = np.array([[2, 2, 2], [2, 4, 5]], np.float)
-    assert compute_min_dist(mol1_pos, mol2_pos, mol3_pos) < 0.1
+    assert pipeline.compute_min_dist(mol1_pos, mol2_pos, mol3_pos) < 0.1
     mol1_pos = remove_overlap(mol1_pos, mol2_pos, mol3_pos, min_distance=0.1, sigma=2.0)
-    assert compute_min_dist(mol1_pos, mol2_pos, mol3_pos) >= 0.1
+    assert pipeline.compute_min_dist(mol1_pos, mol2_pos, mol3_pos) >= 0.1
 
 
 def test_pull_close():
@@ -218,8 +200,8 @@ def test_pull_close():
     translation2 = pull_close(mol1_pos, mol2_pos, 1.5, 5)
     translation3 = pull_close(mol1_pos, mol3_pos, 1.5, 5)
     assert isinstance(translation2, np.ndarray)
-    assert 1.5 <= compute_min_dist(mol1_pos, mol2_pos + translation2) <= 5
-    assert 1.5 <= compute_min_dist(mol1_pos, mol3_pos + translation3) <= 5
+    assert 1.5 <= pipeline.compute_min_dist(mol1_pos, mol2_pos + translation2) <= 5
+    assert 1.5 <= pipeline.compute_min_dist(mol1_pos, mol3_pos + translation3) <= 5
 
 
 def test_pack_transformation():
@@ -236,7 +218,7 @@ def test_pack_transformation():
     for mol, transf in zip(mols_affine, transformations):
         assert isinstance(transf, np.ndarray)
         mol2 = mol.dot(transf.T)[:, :3]  # transform and "de-affine"
-        min_dist, max_dist = compute_min_max_dist(mol1, mol2)
+        min_dist, max_dist = pipeline.compute_min_max_dist(mol1, mol2)
         assert CLASH_DIST <= min_dist and max_dist <= BOX_SIZE
 
 
@@ -770,7 +752,7 @@ def test_clashing_atoms():
         # Sanity check: at the beginning molecules clash
         toluene_pos = utils.get_oe_mol_positions(utils.read_oe_molecule(toluene_path))
         benzene_pos = utils.get_oe_mol_positions(utils.read_oe_molecule(benzene_path))
-        assert compute_min_dist(toluene_pos, benzene_pos) < SetupDatabase.CLASH_THRESHOLD
+        assert pipeline.compute_min_dist(toluene_pos, benzene_pos) < SetupDatabase.CLASH_THRESHOLD
 
         yaml_builder = YamlBuilder(yaml_content)
 
@@ -787,7 +769,7 @@ def test_clashing_atoms():
             toluene_pos2 = positions.take(atom_indices['ligand'], axis=0)
 
             # Test that clashes are resolved in the system
-            min_dist, max_dist = compute_min_max_dist(toluene_pos2, benzene_pos2)
+            min_dist, max_dist = pipeline.compute_min_max_dist(toluene_pos2, benzene_pos2)
             assert min_dist >= SetupDatabase.CLASH_THRESHOLD
 
             # For solvent we check that molecule is within the box
